📝 Add instructions to install for ember
Added some instructions to install this package and use it in an EmberCLI app.

My team had a bit of a hard time figuring out how to set this up a while back in ember. It works great though! 

If there is a better place for these docs please let me know and I'll move them there.